+++ content_editor/__init__.py
@@ -1,5 +1,5 @@
 from __future__ import absolute_import, unicode_literals
 
 
-VERSION = (1, 1, 2)
+VERSION = (1, 1, 3)
 __version__ = '.'.join(map(str, VERSION))